making C++ functions libraries
Creation of src/testsCpp.cpp, src/ade4libCpp.cpp and inst/include/ade4libCpp.h
