+++ src/Makevars
@@ -1,2 +1,2 @@
- PKG_LIBS =  $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
-
+PKG_LIBS =  $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) 
+PKG_CPPFLAGS = -I../inst/include
